requirements.txt
allure-pytest==2.15.0
flask==3.1.2
ruff==0.14.6